requirements.txt
agno
exa-py
python-dotenv
requests
beautifulsoup4